arrays support chars
